CMake improvements
Win32, MSVC: Enable C++ standard conformance mode

Stolen from SolveSpace:

Win32: Allow 32 bit SolveSpace to access up to 4GB of RAM.

Link 32 bit SolveSpace for Windows with /LARGEADDRESSAWARE which allows
it to access up to 3GB of RAM on a properly configured 32 bit Windows and
up to 4GB on 64 bit.

Win32, MSVC: Enable Multi-processor Compilation (/MP) with Visual Studio

Makes compiling from the Visual Studio IDE much faster when using the
solution and projects generated by cmake.

For the externals I hijack the `disable_warnings` function.

Extlib fixups and macOS framework finding fix (#1502)

build: fix CMake system include directives

The CMake `SYSTEM` modifier of `target_include_directories()` must come
before the visibility specification. However, it was accidentally in the
wrong order, so it was simply adding a directory named `SYSTEM` to the
target's include paths.

Fix it to correctly request the include paths to be treated as system
include paths.

build: fix building with OpenMP on macOS

Something changed in the way compiler flags are passed by CMake to the
compiler, which caused the OpenMP flags to be passed as a single flag
and resulting in a build failure.

Use CMake's `SHELL:` syntax to tell CMake to pass these as separate
flags, but without reordering them or performing deduplication.

build: allow enabling ASan on Windows with MSVC

ASan is supported with MSVC starting with Visual Studio 16.9, so we can
allow it if the right version is detected.

The syntax is different (each sanitiser should be provided as a separate
flag), and of the sanitisers that we support, only ASan is available in
MSVC. The new logic handles both of these MSVC-specific changes.

Win32: Make SolveSpace grea...static again

For some reason the /MT options in the `CMAKE_C_FLAGS_RELEASE_INIT` etc.
CMake targets stopped working at some point and SolveSpace became linked to
the DLL versions of the C/C++ runtime (MSVCP140.DLL, VCRUNTIME140.DLL).

We do not like this. Therefore:

Use the proper "modern" CMake incantation to make SolveSpace static again ;-)
https://cmake.org/cmake/help/latest/prop_tgt/MSVC_RUNTIME_LIBRARY.html

The OpenMP build still depends on VCOMP140.DLL but this seems unavoidable.
https://stackoverflow.com/questions/37037733/statically-link-with-microsoft-crt-and-openmp
https://github.com/godotengine/godot/issues/14935